Split parser for sql inside and outside of EXEC SQL
