+++ server/engine/src/main/antlr4/org/eclipse/lsp/cobol/implicitDialects/sql/Db2SqlExecLexer.g4
@@ -11,7 +11,7 @@
  * Contributors:
  *   Broadcom, Inc. - initial API and implementation
  */
-lexer grammar Db2SqlLexer;
+lexer grammar Db2SqlExecLexer;
 channels{COMMENTS}
 
 // DB2 SQL Reserved Keywords
@@ -861,7 +861,14 @@ INTEGERLITERAL : DIGIT+;
 IDENTIFIER : [\p{Alnum}\p{General_Category=Other_Letter}] [-_\p{Alnum}\p{General_Category=Other_Letter}]*;
 COPYBOOK_IDENTIFIER : [a-zA-Z0-9#@$][_a-zA-Z0-9#@$]*;
 
-NUMERICLITERAL : (PLUSCHAR | MINUSCHAR)? DIGIT* (DOT_FS | COMMACHAR) DIGIT+ (('e' | 'E') (PLUSCHAR | MINUSCHAR)? DIGIT+)?;
+NUMERICLITERAL : (PLUSCHAR | MINUSCHAR)? DIGIT*
+    (
+        DIGIT? (DOT_FS | COMMACHAR {_input.LA(1) != ' ' && _input.LA(1) != NEWLINE}?) DIGIT
+        |
+        DIGIT (DOT_FS | COMMACHAR {_input.LA(1) != ' '&& _input.LA(1) != NEWLINE }?) DIGIT?
+    ) DIGIT*
+    (('e' | 'E') (PLUSCHAR | MINUSCHAR)? DIGIT+)?;
+
 NONNUMERICLITERAL : UNTRMSTRINGLITERAL | STRINGLITERAL | DBCSLITERAL | HEXNUMBER | NULLTERMINATED;
 
 CHAR_STRING_CONSTANT : HEXNUMBER | STRINGLITERAL;
