+++ server/engine/src/main/java/org/eclipse/lsp/cobol/implicitDialects/sql/Db2SqlDialect.java
@@ -70,7 +70,7 @@ public class Db2SqlDialect implements CobolDialect {
 
   @Override
   public ResultWithErrors<DialectOutcome> processText(DialectProcessingContext context) {
-    Db2SqlVisitor db2SqlVisitor = new Db2SqlVisitor(context, messageService, copybookService);
+    Db2SqlUtilityVisitor db2SqlVisitor = new Db2SqlUtilityVisitor(context, messageService, copybookService);
 
     List<SyntaxError> parseError = new ArrayList<>();
